backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
decouple==0.0.7
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
fonttools==4.42.1
h11==0.14.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
psycopg==3.1.10
psycopg-binary==3.1.10
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
